+++ packages/button/src/utils/getTestUtils.types.tsx
@@ -1,0 +1,6 @@
+import { FormUtils } from '@lg-tools/test-harnesses';
+
+export interface GetTestUtilsReturnType<T extends HTMLElement> {
+  getButton: () => T;
+  isDisabled: FormUtils['isDisabled'];
+}