+++ packages/skeleton-loader/src/CodeSkeleton/CodeSkeleton.stories.tsx
@@ -1,24 +1,33 @@
 import React from 'react';
-import { storybookArgTypes } from '@lg-tools/storybook-utils';
 import { StoryFn } from '@storybook/react';
+import { StoryMetaType, StoryType} from '@lg-tools/storybook-utils';
 
 import { CodeSkeleton } from '..';
 
 export default {
   title: 'Components/SkeletonLoader',
   component: CodeSkeleton,
-  argTypes: {
-    darkMode: storybookArgTypes.darkMode,
+  parameters: {
+    default: null,
+    controls: { exclude: ['darkMode', 'ref'] },
+    generate: {
+      storyNames: ['Code'],
+      combineArgs: {
+        darkMode: [false, true],
+      },
+      decorator: (Instance: StoryFn) => (
+        <div style={{ width: 500 }}>
+          <Instance />
+        </div>
+      ),
+    },
   },
-  decorators: [
-    (Story: StoryFn) => (
-      <div style={{ width: 700 }}>
-        <Story />
-      </div>
-    ),
-  ],
-};
+  args: {
+    enableAnimations: false,
+  },
+  argTypes: {
+    enableAnimations: { control: 'boolean' },
+  },
+} satisfies StoryMetaType<typeof CodeSkeleton>;
 
-export const Code: StoryFn<typeof CodeSkeleton> = props => (
-  <CodeSkeleton {...props} />
-);
+export const Code: StoryType<typeof CodeSkeleton> = () => <></>;